# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeSci Nexus Phase 1 workspace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  ACTIVE: 'ACTIVE',
  ARCHIVED: 'ARCHIVED',
} as const;

export type WorkflowState = typeof WorkflowState[keyof typeof WorkflowState];


export const WorkflowState = {
  CREATED: 'CREATED',
  SCOPED: 'SCOPED',
  DOCUMENTS_ANALYZED: 'DOCUMENTS_ANALYZED',
  EVIDENCE_RETRIEVED: 'EVIDENCE_RETRIEVED',
  GAPS_IDENTIFIED: 'GAPS_IDENTIFIED',
  FINDINGS_VERIFIED: 'FINDINGS_VERIFIED',
  REPORT_DRAFTED: 'REPORT_DRAFTED',
  HUMAN_REVIEW: 'HUMAN_REVIEW',
  APPROVED: 'APPROVED',
  RETURNED: 'RETURNED',
  REJECTED: 'REJECTED',
} as const;

export type FindingStatus = typeof FindingStatus[keyof typeof FindingStatus];


export const FindingStatus = {
  SUPPORTED: 'SUPPORTED',
  PARTIALLY_SUPPORTED: 'PARTIALLY_SUPPORTED',
  UNSUPPORTED: 'UNSUPPORTED',
  CONFLICTING: 'CONFLICTING',
  NOT_ASSESSABLE: 'NOT_ASSESSABLE',
  REQUIRES_HUMAN_REVIEW: 'REQUIRES_HUMAN_REVIEW',
} as const;

export interface ProjectScope {
  productName: string;
  intendedUse: string;
  projectPhase: string;
  assessmentFramework: string;
  /** @nullable */
  jurisdictionContext?: string | null;
  /** @nullable */
  scopeNotes?: string | null;
}

export interface Project {
  id: string;
  name: string;
  status: ProjectStatus;
  scope: ProjectScope;
  workflowState: WorkflowState;
  findingCount: number;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  scope: ProjectScope;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  status?: ProjectStatus;
  scope?: ProjectScope;
}

export interface Run {
  id: string;
  projectId: string;
  state: WorkflowState;
  label: string;
  startedAt: string;
  updatedAt: string;
}

export interface RunInput {
  /** @minLength 1 */
  label?: string;
}

export type FindingConfidence = typeof FindingConfidence[keyof typeof FindingConfidence];


export const FindingConfidence = {
  HIGH: 'HIGH',
  MEDIUM: 'MEDIUM',
  LOW: 'LOW',
} as const;

export type ProvenanceSourceType = typeof ProvenanceSourceType[keyof typeof ProvenanceSourceType];


export const ProvenanceSourceType = {
  EXTERNAL_EVIDENCE: 'EXTERNAL_EVIDENCE',
  PROJECT_MEMORY: 'PROJECT_MEMORY',
  PROJECT_DOCUMENT: 'PROJECT_DOCUMENT',
  HUMAN_DECISION: 'HUMAN_DECISION',
} as const;

export interface Provenance {
  sourceType: ProvenanceSourceType;
  label: string;
  /** @nullable */
  locator?: string | null;
}

export interface Finding {
  id: string;
  projectId: string;
  title: string;
  summary: string;
  status: FindingStatus;
  confidence: FindingConfidence;
  provenance: Provenance[];
  /** @nullable */
  reviewerDecision?: string | null;
  updatedAt: string;
}

export type ReviewItemPriority = typeof ReviewItemPriority[keyof typeof ReviewItemPriority];


export const ReviewItemPriority = {
  HIGH: 'HIGH',
  MEDIUM: 'MEDIUM',
  LOW: 'LOW',
} as const;

export interface ReviewItem {
  id: string;
  projectId: string;
  title: string;
  reason: string;
  priority: ReviewItemPriority;
  createdAt: string;
}

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  DRAFT: 'DRAFT',
  HUMAN_REVIEW: 'HUMAN_REVIEW',
  APPROVED: 'APPROVED',
  RETURNED: 'RETURNED',
} as const;

export interface Report {
  id: string;
  projectId: string;
  title: string;
  status: ReportStatus;
  generatedAt: string;
  findingCount: number;
}

export interface AuditEvent {
  id: string;
  projectId: string;
  eventType: string;
  actor: string;
  description: string;
  createdAt: string;
}

export interface ProjectWorkspace {
  project: Project;
  runs: Run[];
  findings: Finding[];
  reports: Report[];
  recentAuditEvents: AuditEvent[];
}

export interface Dashboard {
  activeProjects: number;
  reviewQueueCount: number;
  verifiedFindingCount: number;
  recentActivity: AuditEvent[];
}

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

export type ProjectStatusQueryParameter = ProjectStatus;

export type FindingStatusQueryParameter = FindingStatus;

export type ListProjectsParams = {
status?: ProjectStatusQueryParameter;
};