# lib/api-zod/src/generated/types/auditEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeSci Nexus Phase 1 workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditEvent {
  id: string;
  projectId: string;
  eventType: string;
  actor: string;
  description: string;
  createdAt: Date;
}