# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeSci Nexus Phase 1 workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditEvent } from './auditEvent';

export interface Dashboard {
  activeProjects: number;
  reviewQueueCount: number;
  verifiedFindingCount: number;
  recentActivity: AuditEvent[];
}