# lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeSci Nexus Phase 1 workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { FindingConfidence } from './findingConfidence';
import type { FindingStatus } from './findingStatus';
import type { Provenance } from './provenance';

export interface Finding {
  id: string;
  projectId: string;
  title: string;
  summary: string;
  status: FindingStatus;
  confidence: FindingConfidence;
  provenance: Provenance[];
  /** @nullable */
  reviewerDecision?: string | null;
  updatedAt: Date;
}