# lib/api-zod/src/generated/types/findingStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeSci Nexus Phase 1 workspace API
 * OpenAPI spec version: 0.1.0
 */

export type FindingStatus = typeof FindingStatus[keyof typeof FindingStatus];


export const FindingStatus = {
  SUPPORTED: 'SUPPORTED',
  PARTIALLY_SUPPORTED: 'PARTIALLY_SUPPORTED',
  UNSUPPORTED: 'UNSUPPORTED',
  CONFLICTING: 'CONFLICTING',
  NOT_ASSESSABLE: 'NOT_ASSESSABLE',
  REQUIRES_HUMAN_REVIEW: 'REQUIRES_HUMAN_REVIEW',
} as const;