# lib/api-zod/src/generated/types/projectScope.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeSci Nexus Phase 1 workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectScope {
  productName: string;
  intendedUse: string;
  projectPhase: string;
  assessmentFramework: string;
  /** @nullable */
  jurisdictionContext?: string | null;
  /** @nullable */
  scopeNotes?: string | null;
}