# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeSci Nexus Phase 1 workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectScope } from './projectScope';
import type { ProjectStatus } from './projectStatus';

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  status?: ProjectStatus;
  scope?: ProjectScope;
}