# lib/api-zod/src/generated/types/projectWorkspace.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeSci Nexus Phase 1 workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditEvent } from './auditEvent';
import type { Finding } from './finding';
import type { Project } from './project';
import type { Report } from './report';
import type { Run } from './run';

export interface ProjectWorkspace {
  project: Project;
  runs: Run[];
  findings: Finding[];
  reports: Report[];
  recentAuditEvents: AuditEvent[];
}