# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeSci Nexus Phase 1 workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportStatus } from './reportStatus';

export interface Report {
  id: string;
  projectId: string;
  title: string;
  status: ReportStatus;
  generatedAt: Date;
  findingCount: number;
}