# lib/api-zod/src/generated/types/reviewItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeSci Nexus Phase 1 workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewItemPriority } from './reviewItemPriority';

export interface ReviewItem {
  id: string;
  projectId: string;
  title: string;
  reason: string;
  priority: ReviewItemPriority;
  createdAt: Date;
}