# lib/api-zod/src/generated/types/run.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeSci Nexus Phase 1 workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkflowState } from './workflowState';

export interface Run {
  id: string;
  projectId: string;
  state: WorkflowState;
  label: string;
  startedAt: Date;
  updatedAt: Date;
}