# lib/api-zod/src/generated/types/runInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * LifeSci Nexus Phase 1 workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface RunInput {
  /** @minLength 1 */
  label?: string;
}